% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/individ_ic_fun.R
\name{approx_pmp}
\alias{approx_pmp}
\title{Compute individual and model averaged results using the AIC
  and BIC approximations.}
\usage{
approx_pmp(inmat, Xmat, Yvec, mod_names = NULL, mod_prior = NULL,
  family = "gaussian", w_plus = TRUE)
}
\arguments{
\item{inmat}{A matrix with dimesions \eqn{K \times p} specifying which
variables are included in each model.}

\item{Xmat}{A matrix of the potential covariates to consider
(dimension \eqn{n \times p}).}

\item{Yvec}{A vector of responses (dimension \eqn{n \times 1}).}

\item{mod_names}{An optional vector specifying the names of the \eqn{K}
models being considered.}

\item{mod_prior}{A vector of prior probabilities for the set of models
for computing BIC model weights. If \code{NULL}, the prior defaults to
\eqn{\frac{1}{K}} for all \eqn{K} models.}

\item{family}{An input specifying the \eqn{glm} family to use in
estimation and computation of IC criteria.}

\item{w_plus}{A logical statement to pick the version of the formula
to use. LIST FORMULAE here**... Default is \code{TRUE}, which does the
computation over the set of models for which the variable associated
with the coefficient is included.}
}
\value{
A \code{data.frame} with information about the following variables
  for all models in the model set and the model averaged result: Model,
  approximate weights for AIC and BIC, AIC, BIC, and ML coefficinet estimates
  and their associated standard errors. Pieces of this ouptut can be used with
  the \code{MApp_IC} function.
}
\description{
Create a \code{data.frame} summarizing individual and
  model averaged results for partial regression coefficients from
  a set of models.
}

