% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individ_post_fun.R
\name{bms_post_sim}
\alias{bms_post_sim}
\title{Simulate draws from posterior distributions of partial regression
  coefficients.}
\usage{
bms_post_sim(input_mat, Xmat, Yvec, num_sims, g = NULL)
}
\arguments{
\item{input_mat}{A matrix defining the model set.}

\item{Xmat}{The numeric p by n X matrix for the regression using all p
covariates.}

\item{Yvec}{The numeric vector of responses.}

\item{num_sims}{The number of draws to simulate.}
}
\value{
A list of simulated posterior distributions for all partial
  regression coefficients for each model considered.
}
\description{
Simulate draws from individual posterior distributions for all
  models in the model set. Posteriors are simulated assuming a
   g-prior was used on the coefficints.
}
\seealso{
Other MApp.functions: \code{\link{MApost_bms_sim}}
}

