% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/individ_post_fun.R
\name{cov.fun}
\alias{cov.fun}
\title{Compute analytical posterior variance covariance matrix of
  partial regression coefficients}
\usage{
cov.fun(lmObject)
}
\arguments{
\item{lmObject}{An object of class \code{lm}}
}
\value{
The posterior variance for the individual \code{lm} object.
}
\description{
Compute posterior variance covariance matrix
for an individual
regression model after model averaging.
Posterior means are generated assuming
model averaging with the \code{bms()} function,
with a unit information prior (g = N) placed
on the regression parameters.
}

