% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/individ_post_fun.R
\name{cov_fun}
\alias{cov_fun}
\title{Compute analytical posterior variance covariance matrix of partial
  regression coefficients}
\usage{
cov_fun(lmObject, g = NULL, ...)
}
\arguments{
\item{lmObject}{An object of class \code{lm}}
}
\value{
The posterior variance for the individual \code{lm} object.
}
\description{
Compute posterior variance covariance matrix for an individual
  regression model after model averaging. Posterior variance-covariance
  matrices are generated assuming model averaging with Zellner's
  \eqn{g}-prior [add citation here]. If no value of \eqn{g} is specified, the
  unit information prior (g = N) is assumed. placed on the regression
  parameters.
}

