% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/individ_ic_fun.R
\name{est_MA}
\alias{est_MA}
\title{Compute model averaged estimates for partial regression
  coefficinets.}
\usage{
est_MA(weights, coef_mat)
}
\arguments{
\item{weights}{A vector of approximate posterior model probabilties
(commonly referred to as model weights in the IC literature).}

\item{coef_mat}{A \eqn{K} by \eqn{p} matrix of maximum likeilhood
estimates for the partial regression coefficinets from all individual
models considered.}
}
\value{
The point estimate of the model averaged partial regression
  coefficients.
}
\description{
This function computes point estimates for model averaged
  partial regression coefficients using a weighted avearged of maximum
  likelihood estimates from individual models according to approximate
  posterior model probabilities.
}

