% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individ_post_fun.R
\name{MApost_bms_sim}
\alias{MApost_bms_sim}
\title{Generate individual and MA posterior distributions}
\usage{
MApost_bms_sim(x, num_draws = 5000, mod_names = NULL, ...)
}
\arguments{
\item{x}{a \code{bma} object generated by \code{bms()}}

\item{num_draws}{The size of the model averaged posterior distribution.}

\item{mod_names}{a vector of the model names}
}
\value{
\code{list}. The first element is the model averaged posterior 
		distributions for the partial regression coefficients 
	associated with each explanatory variable and the number 
		of samples taken from each model in \eqn{\mathcal{M}}.
}
\description{
Simulate draws from individual posterior distributions for all
  models in the model set. Posteriors are simulated assuming a
   g-prior was used on the coefficients. Use individual posterior draws and
	 posterior model probabilities to sample from the MA posterior distribution.
}
\seealso{
Other MApp.functions: \code{\link{bms_post_sim}}
}

