% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_fun_bms.R
\name{MApp.bms}
\alias{MApp.bms}
\title{Model Averaged Poseriors Plot for bms objects}
\usage{
MApp.bms(x, plot.wind, num.sims = 1000, max.display = NULL,
  mod.names = NULL, include.coef = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{bms} where}

\item{plot.wind}{A vector of length 2 specifying the number of rows and
columns to partition the plotting window into.}

\item{max.display}{An integer specifying the number of models to display in
the plotting window.}

\item{mod.names}{A vector specifying the names of the models. If left blank,
models will be named \eqn{M_1, M_2,..., M_K}, where \eqn{M_1} will have the
largest posterior model probability and \eqn{M_K} has the smallest
posterior model probability.}

\item{include.coef}{A vector specifying which partial regression coefficients
to display.}
}
\value{
The \code{MApp.bms} plot and a table of posterior standard deviations for
  the top \code{max.display} individual models specified, along with the
  posterior standard deviation of the model averaged parameter.
}
\description{
Compare posterior distributions of partial regression
  coefficients from individual models considered to the posterior
  distribuitons of the model graphically and compare poserior standard
  deviations in tabular form.
}

