% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting_fun.R
\name{MApp}
\alias{MApp}
\title{Plots the posterior distributions of a parameter of interest along with the
model averaged posterior distribuiton for the same parameter.}
\usage{
MApp(mcmc.list, weights, PIP, plot.wind, max.display = NULL,
  mod.names = NULL, include.coef = NULL, ...)
}
\arguments{
\item{mcmc.list}{A list of simulated posterior distribuitons for all
partial regression coefficients for each model considered.}

\item{weights}{A vector of posterior model weights for all models in the model
set.}

\item{PIP}{A vector of posterior inclusion probabilities for all potential
covariates.}

\item{plot.wind}{A vector of length 2 specifying the number of rows and columns
to partition the plotting window into.}

\item{max.display}{An integer specifying the number of models to display in
the plotting window.}

\item{mod.names}{A vector specifying the names of the models. If left blank,
models will be named M_1, M_2,..., M_K, where M_1 will have the largest
posterior model probability and M_K has the smallest posterior model
probability.}

\item{include.coef}{A vector specifying which partial regression
coefficients to display.}
}
\value{
A table of posterior standard deviations for the top \code{max.display}
 individual models specified, along with the posterior standard deviation of
 the model averaged parameter.
}
\description{
Plots the posterior distributions of a parameter of interest along with the
model averaged posterior distribuiton for the same parameter.
}

