% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fun_bms.R
\name{MApp_bms}
\alias{MApp_bms}
\title{Model Averaged Posteriors Plot (MAP plot)}
\usage{
MApp_bms(x, plot_wind, num_sims = 1000, max_display = NULL,
  mod_names = NULL, include_coef = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{bma} from the \code{BMS::bms()} function}

\item{plot_wind}{A vector of length 2 specifying the number of rows and
columns to partition the plotting window into.}

\item{num_sims}{the number of Monte Carlo samples to generate from each individual 
posterior distribution for each partial regression coefficient of interest. 
This will also be the number of observations in the model averaged posterior 
distribution. Default is 1000.}

\item{max_display}{An integer specifying the number of models to display in
the plotting window. Can also take on the argument \code{"common3"} to show the
top PMP model, the full model, and the model averaged distribution. If left 
blank, the MAP plot will display posterior distributions from the 20 models
with largest 
posterior model probabilities.}

\item{mod_names}{A vector specifying the names of the models. If left blank,
models will be named \eqn{M_1, M_2,..., M_J}, where \eqn{M_1} will have the
largest posterior model probability and \eqn{M_J} has the smallest
posterior model probability.}

\item{include_coef}{A vector specifying which partial regression coefficients
to display.}
}
\value{
The \code{MApp_bms} plot and a table of posterior standard deviations
  for the top \code{max_display} individual models specified, along with the
  posterior standard deviation of the model averaged parameter.
}
\description{
For \code{bma} objects obtained using the \code{bms} function in
  the \code{BMS} package. Creates the model averaged posteriors plot, displaying 
  posteriors of regression coefficients from individual models along with the 
  corresponding model averaged posteriors. Used to visually display the components 
  of model averaged partial regression coefficients and assess appropriateness of 
  drawing inference from them. Also provides a table of
  posterior standard deviations for the model averaged
  distribution and the distributions from individual models.
}

