% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fun_mcmc_list.R
\name{MApp_list}
\alias{MApp_list}
\title{Model Averaged Posteriors Plot}
\usage{
MApp_list(mcmc_list, plot_wind, max_display = NULL, mod_names = NULL,
  include_coef = NULL, ...)
}
\arguments{
\item{mcmc_list}{A list of posterior samples of partial regression
coefficients of interest for each model in the model set.}

\item{plot_wind}{A vector of length 2 specifying the number of rows and
columns to partition the plotting window into.}

\item{max_display}{An integer specifying the number of models to display in
the plotting window.}

\item{mod_names}{A vector specifying the names of the models. If left blank,
models will be named \eqn{M_1, M_2,..., M_J}.}

\item{include_coef}{A vector specifying which partial regression coefficients
to display.}
}
\value{
MAP plot and a table of posterior standard deviations for
  the top \code{max_display} individual models specified, along with the
  posterior standard deviation from the model averaged result.
}
\description{
Works with lists of posterior draws of partial regression
  coefficients from multiple models. Creates a graphical display showing all
  relevant output from the model averaging procedure. Used to graphically
  compare posterior distributions of partial regression coefficients from
  individual models in the model set to the posterior distributions resulting
  from the model averaged partial regression coefficients. Also used to
  compare posterior standard deviations of the model
  averaged distribution to corresponding standard deviations from individual
  posterior distributions in tabular form.
}

