% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/individ_post_fun.R
\name{mean.fun}
\alias{mean.fun}
\title{Compute analytical posterior means}
\usage{
\method{mean}{fun}(lmObject)
}
\arguments{
\item{lmObject}{A lm object}
}
\value{
The posterior mean for the individual\code{lm} object.
}
\description{
Compute posterior means for an individual regression model after
  model averaging. Posterior means are generated assuming model averaging
  with the \code{bms()} function, with a unit information prior (g = N)
  placed on the regression parameters.
}

