% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/individ_post_fun.R
\name{mean_fun}
\alias{mean_fun}
\title{Compute analytical posterior means}
\usage{
mean_fun(lmObject, g = NULL, ...)
}
\arguments{
\item{lmObject}{A lm object}
}
\value{
The posterior mean for the individual\code{lm} object.
}
\description{
Compute posterior means for an individual regression model after
  model averaging. Posterior means are generated assuming model averaging
  with Zellner's \eqn{g}-prior [add citation here]. If no value of \eqn{g} is
  specified, the unit information prior (g = N) is assumed. placed on the
  regression parameters.
}

