% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individ_ic_fun.R
\name{pmp_AIC}
\alias{pmp_AIC}
\title{Compute approximate posterior model probabilities using the AIC
  approximation for a set of models.}
\usage{
pmp_AIC(AIC_vec)
}
\arguments{
\item{AIC_vec}{A vector of AIC values for a set of models.}
}
\value{
The approximate posterior model probability. (Formulae
according to Link & Barker, 2010; Burhnam and Anderson, 2002)
}

