% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/individ_ic_fun.R
\name{pmp_BIC}
\alias{pmp_BIC}
\title{Compute approximate posterior model probabilities using the BIC
  approximation for a set of models.}
\usage{
pmp_BIC(BIC_vec, mod_prior)
}
\arguments{
\item{BIC_vec}{A vector of BIC values for a set of models.}

\item{mod_prior}{A vector of prior probabilities for the set of models.}
}
\value{
The approximate posterior model probability. (Formulae
according to Link & Barker, 2010; Burhnam and Anderson, 2002)
}
\description{
Compute approximate posterior model probabilities using the BIC
  approximation for a set of models.
}

