% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individ_ic_fun.R
\name{se_MA}
\alias{se_MA}
\title{Compute the standard error for model averaged
  partial regression coefficients.}
\usage{
se_MA(weights, coef_mat, se_mat, inmat, w_plus = FALSE)
}
\arguments{
\item{weights}{A vector of approximate posterior model probabilities
(commonly referred to as model weights in the IC literature).}

\item{coef_mat}{A \eqn{J} by \eqn{p} matrix of maximum likelihood
estimates for the partial regression coefficients from all individual
models considered.}

\item{se_mat}{A \eqn{J} by \eqn{p} matrix of maximum likelihood estimates
of standard errors for the partial regression coefficients from all
individual models considered.}

\item{inmat}{A matrix with dimensions \eqn{J \times p} specifying which
variables are included in each model. The order of models must correspond
to the order of \code{coef_mat} a \code{ses_mat} (it is recommended, but not
required to sort these in decreasing posterior model probability).}

\item{w_plus}{A logical statement defining which type of MA to use. 
Default is \code{FALSE}, which allows all models defined in \code{inmat} to 
be used in the averaging for each variable. If set to \code{TRUE}, MA
estimates will be made using only the models for which the coefficient
associated with each variable is not set to exactly zero. 
Weights are normalized conditional on the models meeting this criterion.}
}
\value{
The point estimate of the model averaged partial regression
  coefficients.
}
\description{
This function computes standard errors for model averaged
  partial regression coefficients using the formula specified in
  Link and Barker 2010 (see also Burhnam and Anderson (2002)).
}

