% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individ_post_fun.R
\name{vec_fun}
\alias{vec_fun}
\title{Convert a list of posterior samples into a long data frame.}
\usage{
vec_fun(post.list, coef)
}
\description{
This function will create a long data frame to be used as an
  input in \code{\link{MApost_bms_sim}}.
}

