% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_ex_data.R
\docType{data}
\name{brainData}
\alias{brainData}
\title{Brain weight, body weight, gestation length and litter size}
\format{A data frame with 96 rows and 4 variables: \describe{
  \item{lbrain}{log transformed average brain weight, measured in log(grams)}
  \item{lbody}{log transformed average body weight, measured in log(kg)}
  \item{lgest}{log transformed average gestation length, measured in log(days)}
  \item{litter}{average number of offspring in one litter measured as # of
   offspring} }}
\source{
Original by F.L. Ramsey and D.W. Schafer and modifications by Daniel
  W. Schafer and Jeannie Sifneos and Berwin A. Turlach, (2012). Sleuth2: Data
  sets from Ramsey and Schafer's "Statistical Sleuth (2nd ed)",
  \url{http://CRAN.R-project.org/package=Sleuth2}
}
\usage{
brainData
}
\description{
A dataset containing the log brain weight log(g), log body weight log(kg),
log gestation length log(days), and litter size (# offspring) for 96 species
of mammals. These data were modified from \code{case0902} in the R package 
\code{Sleuth2}.
}
\keyword{datasets}

