% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_fun_mcmc_list.R
\name{MApp}
\alias{MApp}
\title{Model Averaged Posteriors Plot}
\usage{
MApp(mcmc_list, plot_wind, max_display = NULL, mod_names = NULL,
  include_coef = NULL, ...)
}
\arguments{
\item{mcmc.list}{A list of posterior samples of partial regression
coefficients of interest for each model in the model set.}

\item{g}{A numeric value specifying the value of the hyperparameter \eqn{g}.}

\item{weights}{A vector of posterior model weights for all models in the
model set.}

\item{PIP}{A vector of posterior inclusion probabilities for all potential
covariates.}

\item{plot.wind}{A vector of length 2 specifying the number of rows and
columns to partition the plotting window into.}

\item{max.display}{An integer specifying the number of models to display in
the plotting window.}

\item{mod.names}{A vector specifying the names of the models. If left blank,
models will be named \eqn{M_1, M_2,..., M_J}, where \eqn{M_1} will have the
largest posterior model probability and \eqn{M_J} has the smallest
posterior model probability.}

\item{include.coef}{A vector specifying which partial regression coefficients
to display.}
}
\value{
The \code{MApp} plot and a table of posterior standard deviations for
  the top \code{max.display} individual models specified, along with the
  posterior standard deviation of the model averaged parameter.
}
\description{
Works with lists of posterior draws of partial regression
  coefficients from multiple models. Creates a graphical display showing all
  relevant output from the model averaging procedure. Used to graphically
  compare posterior distributions of partial regression coefficients from
  individual models in the model set to the posterior distributions resulting
  from the model averaged partial regression coefficients. Also used to
  compare posterior standard deviations of the continuous piece of the model
  averaged distribution to corresponding standard deviations from individual
  posterior distributions in tabular form.
}

