% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/individ_ic_fun.R
\name{se_MA}
\alias{se_MA}
\title{Compute the standard error for model averaged
  partial regression coefficients.}
\usage{
se_MA(weights, coef_mat, ses_mat, inmat, w_plus = TRUE)
}
\arguments{
\item{weights}{A vector of approximate posterior model probabilties
(commonly referred to as model weights in the IC literature).}

\item{coef_mat}{A \eqn{K} by \eqn{p} matrix of maximum likeilhood
estimates for the partial regression coefficinets from all individual
models considered.}

\item{ses_mat}{A \eqn{K} by \eqn{p} matrix of maximum likelihood estimates
of standard errors for the partial regression coefficients from all
individual models considered.}

\item{inmat}{A matrix with dimesions \eqn{k \times p} specifying which
variables are included in each model. The order of models must correspond
to the order of \code{coef_mat} a \code{ses_mat} (it is recommended, but not
required to sort these in decreasing posterior model probability).}

\item{w_plus}{A logical statement to pick the version of the formula
to use. LIST FORMULAE here**... Default is \code{TRUE}, which does the
computation over the set of models for which the variable associated
with the coefficient is included.}
}
\value{
The point estimate of the model averaged partial regression
  coefficients.
}
\description{
This function computes standard errors for model averaged
  partial regression coefficients using the formula specfied in
  Link and Barker 2010 (see also Burhnam and Anderson (2002)).
}

